しょぼいカレンダーにアクセスする箇所を外部ファイル化しました
